requirements.txt
absl-py==1.0.0
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
autopep8==1.6.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
Django==4.0.2
django-base64field==1.0
django-cors-headers==3.11.0
django-rest-knox==4.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
entrypoints==0.4
executing==0.8.3
Flask==2.0.2
flatbuffers==2.0
fonttools==4.31.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.9.1
ipython==8.1.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.27
protobuf==3.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.14.0
zipp==3.8.0
